+++ TP01/src/gps/GPSEngine.java
@@ -109,7 +109,7 @@ public abstract class GPSEngine {
 		return true;
 	}
 
-	private  boolean checkOpenAndClosed(Integer cost, GPSState state) {
+	private  boolean checkOpenAndClosed(float cost, GPSState state) {
 		for (GPSNode openNode : open) {
 			if (openNode.getState().compare(state) && openNode.getCost() < cost) {
 				return true;
